fixes for student lifecycle APIs & swagger addition

+++ back-end/src/controller/student-controller.ts
@@ -10,6 +10,18 @@ export class StudentController {
     return this.studentRepository.find()
   }
 
+  async getStudent(request: Request, response: Response, next: NextFunction) {
+      const student = await this.studentRepository.findOne(request.params.id)
+      if(student) {
+        return student
+      } else {
+        response.status(400)
+        return {
+          "message": `User with ID '${request.params.id}' not found.`
+        }
+      }
+  }
+
   async createStudent(request: Request, response: Response, next: NextFunction) {
     const { body: params } = request
 
@@ -21,27 +33,41 @@ export class StudentController {
     const student = new Student()
     student.prepareToCreate(createStudentInput)
 
+    response.status(201)
     return this.studentRepository.save(student)
   }
 
   async updateStudent(request: Request, response: Response, next: NextFunction) {
     const { body: params } = request
 
-    this.studentRepository.findOne(params.id).then((student) => {
+    const studentToUpdate = await this.studentRepository.findOne(params.id)
+    if(studentToUpdate) {
       const updateStudentInput: UpdateStudentInput = {
         id: params.id,
         first_name: params.first_name,
         last_name: params.last_name,
         photo_url: params.photo_url,
       }
-      student.prepareToUpdate(updateStudentInput)
+      studentToUpdate.prepareToUpdate(updateStudentInput)
 
-      return this.studentRepository.save(student)
-    })
+      return this.studentRepository.save(studentToUpdate)
+    } else {
+      response.status(400)
+      return {
+        "message": `User with ID '${params.id}' not found.`
+      }
+    }
   }
 
   async removeStudent(request: Request, response: Response, next: NextFunction) {
     let studentToRemove = await this.studentRepository.findOne(request.params.id)
-    await this.studentRepository.remove(studentToRemove)
+    if(studentToRemove) {
+      return this.studentRepository.remove(studentToRemove)
+    } else {
+      response.status(400)
+      return {
+        "message": `User with ID '${request.params.id}' not found.`
+      }
+    }
   }
 }
